Papel e Permissoes

+++ app/Http/Controllers/PapelController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Papel;
+use App\Models\Permissao;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PapelController extends Controller
 {
@@ -14,6 +17,10 @@ class PapelController extends Controller
      */
     public function index()
     {
+        if(Gate::denies('user-view')){
+            abort(403,"Não autorizado!");
+        }
+
         $registros = Papel::all();
 
         return view('papel.index', compact('registros'));
@@ -26,6 +33,10 @@ class PapelController extends Controller
      */
     public function create()
     {
+        if(Gate::denies('user-create')){
+            abort(403,"Não autorizado!");
+        }
+
         view('papel.create');
     }
 
@@ -37,6 +48,10 @@ class PapelController extends Controller
      */
     public function store(Request $request)
     {
+        if(Gate::denies('usuario-create')){
+            abort(403,"Não autorizado!");
+        }
+
         if ($request->nome && $request->nome != "Admin") {
             Papel::create($request->all());
 
@@ -65,6 +80,10 @@ class PapelController extends Controller
      */
     public function edit($id)
     {
+        if(Gate::denies('user-edit')){
+            abort(403,"Não autorizado!");
+        }
+
         if (Papel::find($id)->nome == "Admin") {
             return redirect()->route('papeis.index');
         }
@@ -83,6 +102,10 @@ class PapelController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(Gate::denies('user-edit')){
+            abort(403,"Não autorizado!");
+        }
+
         if (Papel::find($id)->nome == "Admin") {
             return redirect()->route('papeis.index');
         }
@@ -102,11 +125,42 @@ class PapelController extends Controller
      */
     public function destroy($id)
     {
+        if(Gate::denies('user-destroy')){
+            abort(403,"Não autorizado!");
+        }
+
         if (Papel::find($id)->nome == "Admin") {
             return redirect()->route('papeis.index');
         }
 
         Papel::find($id)->delete();
         return redirect()->route('papeis.index');
+    }
+
+    public function permissao($id)
+    {
+        $papel = Papel::find($id);
+        $permissoes = Permissao::all();
+
+        return view('papel.permissao', compact('papel', 'permissoes'));
+    }
+
+    public function permissaoStore(Request $request, $id)
+    {
+        $papel = Papel::find($id);
+        $data = $request->all();
+        $permissao = Permissao::find($data['permissao_id']);
+        $papel->addPermissao($permissao);
+
+        return redirect()->back();
+    }
+
+    public function permissaoDestroy($id, $permissao_id)
+    {
+        $papel = Papel::find($id);
+        $permissao = Permissao::find($permissao_id);
+        $papel->destroyPermissao($permissao);
+
+        return redirect()->back();
     }
 }
